Make matchHeard() more flexible.
Pass through the heard string with its original encoding.  Tries first
to simple match the input as-is, and only then does it try converting it
to ASCII for a simple match.

Stop trying so hard to match with ASCII string variants -- we can use
fuzzywuzzy for that.

These things both simplify the code here and in the caller, and should
increase match reliability, especially for non-English media.