Remove background cache object refresh.
AWS Lambda doesn't allow worker threads to continue after the main
thread has terminated, so by default, disable this.

We could spawn a separate process for this, but after some discussion we
determined that there just isn't a lot of value in the background
refreshes anyway.

The option to use it remains (config var: cache_bg_update), but is
undocumented.